Extra tailwind plugin into multiple files for better readability
